+++ packages/nimble-components/src/table/tests/table.pageobject.ts
@@ -9,33 +9,36 @@ export class TablePageObject<T extends TableRecord> {
     public constructor(private readonly tableElement: Table<T>) {}
 
     public getRenderedHeaderCount(): number {
-        const header = this.tableElement.shadowRoot!.querySelector('.table-header')!;
-        const cells = header.querySelectorAll('.table-cell');
-        return cells.length;
+        const headers = this.tableElement.shadowRoot!.querySelectorAll(
+            'nimble-table-header'
+        )!;
+        return headers.length;
     }
 
     public getRenderedHeaderContent(columnIndex: number): string {
-        const header = this.tableElement.shadowRoot!.querySelector('.table-header')!;
-        const cells = header.querySelectorAll('.table-cell');
-        if (columnIndex >= cells.length) {
+        const headers = this.tableElement.shadowRoot!.querySelectorAll(
+            'nimble-table-header'
+        )!;
+        if (columnIndex >= headers.length) {
             throw new Error(
                 'Attempting to index past the total number of rendered columns'
             );
         }
 
-        return cells.item(columnIndex).textContent ?? '';
+        return headers.item(columnIndex).textContent?.trim() ?? '';
     }
 
     public getRenderedRowCount(): number {
-        return this.tableElement.shadowRoot!.querySelectorAll('.table-row')
-            .length;
+        return this.tableElement.shadowRoot!.querySelectorAll(
+            'nimble-table-row'
+        ).length;
     }
 
     public getRenderedCellContent(
         rowIndex: number,
         columnIndex: number
     ): string {
-        const rows = this.tableElement.shadowRoot!.querySelectorAll('.table-row');
+        const rows = this.tableElement.shadowRoot!.querySelectorAll('nimble-table-row');
         if (rowIndex >= rows.length) {
             throw new Error(
                 'Attempting to index past the total number of rendered rows'
@@ -43,13 +46,13 @@ export class TablePageObject<T extends TableRecord> {
         }
 
         const row = rows.item(rowIndex);
-        const cells = row.querySelectorAll('.table-cell');
+        const cells = row.shadowRoot!.querySelectorAll('nimble-table-cell');
         if (columnIndex >= cells.length) {
             throw new Error(
                 'Attempting to index past the total number of rendered columns'
             );
         }
 
-        return cells.item(columnIndex).textContent ?? '';
+        return cells.item(columnIndex).shadowRoot!.textContent?.trim() ?? '';
     }
 }